Implemented using Auth0 token passed in request headers to perform logic
We this point we were able to authenticate with Auth0 as well as checking to make sure the user is authorised by verifying the token in our gateway. So, front end gets the token and pass them as headers for our apis to use

I create  a function that extracts the userId out of the auth token and replace the same block of code inside the auth0Authorize with this function.
This function getToken function is created for better documentation as well as avoiding technical debt. Other lambda methods can use it well

I edited the getTodos, creatTodo, updateTodo and deleteTodo to use the userID as well.

I this commit, i also implemented indexing for our DynamoDb tables

More on DynaoDb below

Overview of Dynamdb concepts
https://www.youtube.com/watch?v=Tb8Cw3TUDRM

When and when no to use DynamoDb, the business side
https://www.youtube.com/watch?v=1WF_wped808

Single-Table Design tips with DynamoDB with real world examples
https://www.youtube.com/watch?v=BnDKD_Zv0og
https://www.youtube.com/watch?v=5hQSrWaErK8
https://aws.amazon.com/blogs/database/using-sort-keys-to-organize-data-in-amazon-dynamodb/
https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/GSI.html

DynamoDb streams
https://www.youtube.com/watch?v=_D2Uiufn7NA 